+++ src/domain/data.ts
@@ -12,7 +12,13 @@ import { MdDesignServices } from "react-icons/md";
 import { RiMotorbikeFill } from "react-icons/ri";
 import { SiWebflow, SiZenn } from "react-icons/si";
 import { WtdIcon, WtdIconForResume } from "../common/components/WtdIcon";
-import { AboutData, Data, HeaderDataType, ResumeData } from "./type";
+import {
+  AboutData,
+  CATEGORY,
+  HeaderDataType,
+  PortfolioData,
+  ResumeData,
+} from "./type";
 
 // -------------------
 // Header
@@ -91,7 +97,6 @@ export const aboutData: AboutData = {
 // -------------------
 // Resume
 // -------------------
-
 export const resumeData: ResumeData = {
   history: [
     {
@@ -175,4 +180,86 @@ export const resumeData: ResumeData = {
       },
     ],
   },
+};
+
+// -------------------
+// Portfolio
+// -------------------
+export const portfolioData: PortfolioData = {
+  categoryData: [
+    {
+      name: "all",
+      id: -1,
+    },
+    {
+      name: "coding",
+      id: CATEGORY["CODING"],
+    },
+    {
+      name: "design",
+      id: CATEGORY["DESIGN"],
+    },
+    {
+      name: "logo",
+      id: CATEGORY["LOGO"],
+    },
+  ],
+  mediaItems: [
+    {
+      id: 0,
+      title: "personal cv project",
+      categoryID: CATEGORY["DESIGN"],
+      thumbnailUrl:
+        "https://themes.pixelwars.org/cvcard-wp/wp-content/uploads/2014/04/single-05.png",
+      imgUrl:
+        "https://themes.pixelwars.org/cvcard-wp/wp-content/uploads/2014/04/single-05.png",
+      description:
+        "It is a long established fact that a reader will be distracted by the readable content of a page when looking at its layout. The point of using Lorem Ipsum is that it has a more-or-less normal distribution of letters, as opposed to using.",
+      devStyle: 0,
+      tags: ["HTML5", "CSS3", "TypeScript", "Ruby"],
+      externalUrl: "https://github.com/igsr5",
+    },
+    {
+      id: 1,
+      title: "personal cv project",
+      categoryID: CATEGORY["DESIGN"],
+      thumbnailUrl:
+        "https://themes.pixelwars.org/cvcard-wp/wp-content/uploads/2014/04/single-05.png",
+      imgUrl:
+        "https://themes.pixelwars.org/cvcard-wp/wp-content/uploads/2014/04/single-05.png",
+      description:
+        "It is a long established fact that a reader will be distracted by the readable content of a page when looking at its layout. The point of using Lorem Ipsum is that it has a more-or-less normal distribution of letters, as opposed to using.",
+      devStyle: 1,
+      tags: ["HTML5", "CSS3", "TypeScript", "Ruby"],
+      externalUrl: "https://github.com/igsr5",
+    },
+    {
+      id: 2,
+      title: "personal cv project",
+      categoryID: CATEGORY["DESIGN"],
+      thumbnailUrl:
+        "https://themes.pixelwars.org/cvcard-wp/wp-content/uploads/2014/04/single-05.png",
+      imgUrl:
+        "https://themes.pixelwars.org/cvcard-wp/wp-content/uploads/2014/04/single-05.png",
+      description:
+        "It is a long established fact that a reader will be distracted by the readable content of a page when looking at its layout. The point of using Lorem Ipsum is that it has a more-or-less normal distribution of letters, as opposed to using.",
+      devStyle: 1,
+      tags: ["HTML5", "CSS3", "TypeScript", "Ruby"],
+      externalUrl: "https://github.com/igsr5",
+    },
+    {
+      id: 3,
+      title: "personal cv project",
+      categoryID: CATEGORY["DESIGN"],
+      thumbnailUrl:
+        "https://themes.pixelwars.org/cvcard-wp/wp-content/uploads/2014/04/single-05.png",
+      imgUrl:
+        "https://themes.pixelwars.org/cvcard-wp/wp-content/uploads/2014/04/single-05.png",
+      description:
+        "It is a long established fact that a reader will be distracted by the readable content of a page when looking at its layout. The point of using Lorem Ipsum is that it has a more-or-less normal distribution of letters, as opposed to using.",
+      devStyle: 0,
+      tags: ["HTML5", "CSS3", "TypeScript", "Ruby"],
+      externalUrl: "https://github.com/igsr5",
+    },
+  ],
 };
