+++ src/portfolio/components/MediaItem.tsx
@@ -1,9 +1,9 @@
 import React, { forwardRef } from "react";
+import { CategoryDataType, MediaItemType } from "../../domain/type";
 
-import { MediaItemType } from "./Portfolio";
 import { useRenderMediaItem } from "../hooks/useRenderMediaItem";
 
-type Props = { mediaItem: MediaItemType };
+type Props = { mediaItem: MediaItemType; categoryData: CategoryDataType[] };
 
 export const MediaItem: React.FC<Props> = forwardRef<HTMLDivElement, Props>(
   useRenderMediaItem
